Refactor: rankImg 컴포넌트 추가
