+++ src/components/RankImg.js
@@ -1,0 +1,23 @@
+import React from "react";
+import tier_SS from "../images/tier_SS.png";
+import tier_S from "../images/tier_S.png";
+import tier_A from "../images/tier_A.png";
+import tier_B from "../images/tier_B.png";
+import tier_Un from "../images/tier_UN.png";
+
+function rankImg(tier) {
+    switch (tier) {
+        case "SS":
+            return <img src={tier_SS} alt="tier"></img>;
+        case "S":
+            return <img src={tier_S} alt="tier"></img>;
+        case "A":
+            return <img src={tier_A} alt="tier"></img>;
+        case "B":
+            return <img src={tier_B} alt="tier"></img>;
+        default:
+            return <img src={tier_Un} alt="tier"></img>;
+    }
+}
+
+export default rankImg;
