+++ src/pages/MileagePage.js
@@ -4,14 +4,8 @@ import MainHeader from "../components/MainHeader.js";
 import axios from "axios";
 import PopUp from "../components/PopUp.js";
 import Footer from "../components/Footer.js";
-
-import tier_SS from "../images/tier_SS.png";
-import tier_S from "../images/tier_S.png";
-import tier_A from "../images/tier_A.png";
-import tier_B from "../images/tier_B.png";
-import tier_Un from "../images/tier_UN.png";
+import rankImg from "../components/RankImg.js";
 import MileageHistoryBox from "../components/MileageHistoryBox.js";
-
 import { Container } from "../styles/MileagePageStyled.js";
 
 function MileagePage() {
@@ -77,7 +71,7 @@ function MileagePage() {
     async function postMileage() {
         console.log(addMileage);
         try {
-            const response = await axios.post(
+            await axios.post(
                 "https://api.skhuming-api.store/api/user/mileage/post",
                 {
                     memberId: window.localStorage.getItem("memberId"),
@@ -131,22 +125,6 @@ function MileagePage() {
         getMileageList();
         getMileageHistory();
     }, []);
-
-    // 티어 사진
-    function rankImg(tier) {
-        switch (tier) {
-            case "SS":
-                return <img src={tier_SS} alt="tier"></img>;
-            case "S":
-                return <img src={tier_S} alt="tier"></img>;
-            case "A":
-                return <img src={tier_A} alt="tier"></img>;
-            case "B":
-                return <img src={tier_B} alt="tier"></img>;
-            default:
-                return <img src={tier_Un} alt="tier"></img>;
-        }
-    }
 
     return (
         <Desktop>
