+++ src/pages/MyPage.js
@@ -5,19 +5,12 @@ import MainHeader from "../components/MainHeader.js";
 import ScrapBoxWidget from "../components/ScrapBox_widget.js";
 import axios from "axios";
 import PopUp from "../components/PopUp.js";
-
-import tier_SS from "../images/tier_SS.png";
-import tier_S from "../images/tier_S.png";
-import tier_A from "../images/tier_A.png";
-import tier_B from "../images/tier_B.png";
-import tier_Un from "../images/tier_UN.png";
-
+import rankImg from "../components/RankImg.js";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import {
     faRightFromBracket,
     faSquarePlus,
 } from "@fortawesome/free-solid-svg-icons";
-
 import { Container, StyledLink } from "../styles/MyPageStyled.js";
 
 function MyPage() {
@@ -123,22 +116,6 @@ function MyPage() {
         getScrapData();
         getMileageData();
     }, []);
-
-    // 티어 사진
-    function rankImg(tier) {
-        switch (tier) {
-            case "SS":
-                return <img src={tier_SS} alt="tier"></img>;
-            case "S":
-                return <img src={tier_S} alt="tier"></img>;
-            case "A":
-                return <img src={tier_A} alt="tier"></img>;
-            case "B":
-                return <img src={tier_B} alt="tier"></img>;
-            default:
-                return <img src={tier_Un} alt="tier"></img>;
-        }
-    }
 
     return (
         <Desktop>
