+++ src/pages/NoticeDetailsPage.js
@@ -73,7 +73,6 @@ function NoticeDetailsPage() {
             setMsg(`스크랩 되었습니다!\n마이페이지에서 확인해보세요.`);
             setPopup(true);
         } catch (error) {
-            console.log(error);
             if (error.response.status === 401) {
                 setMsg(error.response.data);
                 setGoLogin(true);
